pyproject.toml
[project]
name = "codedeepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.89.0",
    "openai>=2.30.0",
    "flask>=3.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]
